+++ src/Indice.Features.Cases.AspNetCore/CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [Unreleashed]
+### Bugfix
+- `HttpEndpointWithValidation` : Ignore validation if JSON Schema is null. Null check HTTP Body if there is a Schema to validate it with.
 
 ## [7.21.1] - 2024-03-13
 ### Added